Made booster command add booster and rewrote multiplier to be a double

+++ api/src/main/java/com/alttd/boosterapi/Booster.java
@@ -12,9 +12,9 @@ public interface Booster {
 
     void setType(BoosterType boosterType);
 
-    int getMultiplier();
+    double getMultiplier();
 
-    void setMultiplier(int multiplier);
+    void setMultiplier(double multiplier);
 
     Long getStartingTime();
