+++ velocity/src/main/java/com/alttd/vboosters/data/VelocityBooster.java
@@ -12,11 +12,11 @@ public class VelocityBooster implements Booster {
     private Long startingTime;
     private long duration;
     private BoosterType boosterType;
-    private Integer multiplier;
+    private Double multiplier;
     private Boolean active;
     private Boolean finished;
 
-    public VelocityBooster(UUID uuid, BoosterType boosterType, String reason, long duration, int multiplier) {
+    public VelocityBooster(UUID uuid, BoosterType boosterType, String reason, long duration, double multiplier) {
         this.uuid = uuid;
         this.boosterType = boosterType;
         this.activator = reason;
@@ -27,7 +27,7 @@ public class VelocityBooster implements Booster {
         saveBooster();
     }
 
-    public VelocityBooster(BoosterType type, String playerName, long duration, int multiplier) {
+    public VelocityBooster(BoosterType type, String playerName, long duration, double multiplier) {
         this(UUID.randomUUID(), type, playerName, duration, multiplier);
     }
 
@@ -52,12 +52,12 @@ public class VelocityBooster implements Booster {
     }
 
     @Override
-    public int getMultiplier() {
+    public double getMultiplier() {
         return multiplier;
     }
 
     @Override
-    public void setMultiplier(int multiplier) {
+    public void setMultiplier(double multiplier) {
         this.multiplier = multiplier;
     }
